Fix for null monoterpenes from MEGAN
This is a workaround of an issue in MEGAN. The total monoterpenes
diagnostics is not functioning in the standalone HEMCO and
effectively results in zero total monoterpene emissions.
I've modified GOCART to use the MTPA (a-pinene, b-pinene,
sabinene, carene), MTPO (myrcene, ocimene, other monoterps)
and LIMO (limonene). HEMCO resource files were further trimmed
down.