Fixes #77. Use tiny() in WetRemoval
(This is a re-make of a PR that was merged (#78) and then reverted (#86) to keep a `develop` -> `master` merge clean and zero-diff)

I believe this is the fix for #77. It lets the model run, at least, with GCC at C90. My guess is that the stock GCC Release flags are flushing a verrrrry small number to zero or it's just verrrrrry small. So the if statement says, yes, this is greater than zero, and then the division is getting either a divide-by-zero or a divide-by-denormal. (@tclune might have more insight)

Weirdly, I tested this bugfix with Intel and oddly it was zero-diff at C90. But I am in no way confident this can always be zero-diff since there could be a weird column out there that would trigger a different path, so I'll label it non-zero-diff for now.