Fixes #159. Fix duplicate module builds
Due to overly-aggressive file globs, duplicate `hco_julday_mod` and
`julday_mod` modules are built. This causes an error with the Ninja
build system.

To fix this, we tell our CMake not to build one of the files in HEMCO
and GEOS-Chem. For HEMCO, the files are nigh-identical, so we choose to
omit the one in NcdfUtil for the one in GeosUtil. For GEOS-Chem, we do
the same, but the one in GeosUtil actually seems to have a bugfix.